requirements.txt
streamlit==1.37.1
openai==1.55.3
DateTime==5.5
pandas==2.2.2
python-dotenv==1.0.0
langchain==0.3.9
langchain-community==0.3.9
pysqlite3-binary